RMassBank Representation of an MSMS Spectrum

This extends the Spectrum2 class of the MSnbase package and introduces further slots that are used to store information during the RMassBank workflow.

Slots

satellite

logical If TRUE, the corresponding peak was removed as satellite.

low

logical If TRUE, the corresponding peak was removed because it failed the intensity cutoff.

rawOK

logical If TRUE, the peak passed satellite and low-intensity cutoff removal.

good

logical If TRUE, a formula could be found for the peak and the peak passed all filter criteria. (see the RMassBank vignette or the documentation of analyzeMsMs#' for details on filter settings)

mzCalc

numeric The mz value calculated from the found formula for each peak (if any)

formula

character The formula found for each peak. generate.formula is used for formula-fitting

dbe

numeric The number of double bond equivalents. This is calculated from the found formula for each peak (if any)

formulaCount

integer The number of different formulae found for each peak. Note: A peak for which multiple formulas were found will appear multiple times. Hence there may be multiple entries in the formula , dppm and mzCalc slot for the same mz value.

formulaSource

character "analyze" or "reanalysis" Shows whether the current formula for the peak was determined by normal analysis ("analyze") or by reanalysis of a failpeak ("reanalysis")

dppm

numeric The ppm deviation of the mz value from the found formula (if any).

dppmBest

numeric The ppm deviation of the mz value from the best formula found.

ok

logical one-element vector If this is TRUE, the spectrum was successfully processed with at least one resulting peak. Otherwise, one of the following cases applies:

All peaks failed the intensity cutoff i.e. the whole spectrum contains low intensity peaks, only.

All peaks were marked as satellites.

All peaks in the spectrum have a lower intensity than the value given in the specOkLimit filter setting. (see the RMassBank vignette or the documentation of analyzeMsMs)

The precursor ion formula is invalid (see is.valid.formula)

The spectrum is empty.

No molecular formula could be found for any of the peaks.

All peaks failed the dbeMinLimit criterion. (see the RMassBank vignette or the documentation of analyzeMsMs)

info

list Spectrum identifying information (collision energy, resolution, collision mode) from the spectraList

properties

data.frame This is used as a flexible placeholder to store additional properties for each peak throughout the workflow. After the last step of the mbWorkflow, this will typically contain columns mzRaw, noise, formulaMultiplicity, bestMultiplicity and filterOK. However, new columns may be added on demand (see property<-)

See also