Addition of manual peaklists

Adds a manual peaklist in matrix-format

addPeaksManually(w, cpdID, handSpec, mode)

Arguments

w

The msmsWorkspace that the peaklist should be added to.

cpdID

The compoundID of the compound that has been used for the peaklist

handSpec

A peaklist with 2 columns, one with "mz", one with "int"

mode

The ionization mode that has been used for the spectrum represented by the peaklist

Value

The msmsWorkspace with the additional peaklist added to the right spectrum

See also

Author

Erik Mueller

Examples

if (FALSE) {
    handSpec <- cbind(mz=c(274.986685367956, 259.012401087427, 95.9493025990907, 96.9573002472772),
                               int=c(357,761, 2821, 3446))
    addPeaksManually(w, cpdID, handSpec)
}