Checks for isotopes in a msmsWorkspace

Checks for isotopes in a msmsWorkspace

checkIsotopes(
  w,
  mode = "pH",
  intensity_cutoff = 0,
  intensity_precision = "none",
  conflict = "strict",
  isolationWindow = 2,
  evalMode = "complete",
  plotSpectrum = TRUE,
  settings = getOption("RMassBank")
)

Arguments

w

A msmsWorkspace to work with.

mode

"pH", "pNa", "pM", "pNH4", "mH", "mM", "mFA" for different ions ([M+H]+, [M+Na]+, [M]+, [M+NH4]+, [M-H]-, [M]-, [M+FA]-).

intensity_cutoff

The cutoff (as an absolute intensity value) under which isotopic peaks shouldn't be checked for or accepted as valid. Please note: The cutoff is not hard in the sense that it interacts with the intensity_precision parameter.

intensity_precision

The difference that is accepted between the calculated and observed intensity of a possible isotopic peak. Further details down below.

conflict

Either "isotopic"(Peak formulas are always chosen if they fit the requirements for an isotopic peak) or "strict"(Peaks are only marked as isotopic when there hasn't been a formula assigned before.)

isolationWindow

Half of the width of the isolation window in Da

evalMode

Currently no function yet, but planned. Currently must be "complete"

plotSpectrum

A boolean specifiying whether the spectrumshould be plotted

settings

Options to be used for processing. Defaults to the options loaded via loadRmbSettings et al. Refer to there for specific settings.

Value

The msmsWorkspace with annotated isolation peaks

Details

text describing parameter inputs in more detail.

intensity_precisionThis parameter determines how strict the intensity values should adhere to the calculated intensity in relation to the parent peak. Options for this parameter are "none", where the intensity is irrelevant, "low", which has an error margin of 70% and "high", where the error margin is set to 35%. The recommended setting is "low", but can be changed to adjust to the intensity precision of the mass spectrometer.

Author

Michael Stravs, Eawag <michael.stravs@eawag.ch>

Erik Mueller, UFZ