Extract an MS/MS spectrum from MS2 TIC

Extract an MS/MS spectrum or multiple MS/MS spectra based on the TIC of the MS2 and precursor mass, picking the most intense MS2 scan. Can be used, for example, to get a suitable MS2 from direct infusion data which was collected with purely targeted MS2 without MS1.

findMsMsHR.ticms2(
  msRaw,
  mz,
  limit.coarse,
  limit.fine,
  rtLimits = NA,
  maxCount = NA,
  headerCache = NULL,
  fillPrecursorScan = FALSE,
  deprofile = getOption("RMassBank")$deprofile,
  trace = "ms2tic"
)

Arguments

msRaw

The mzR raw file

mz

Mass to find

limit.coarse

Allowed mass deviation for scan precursor (in m/z values)

limit.fine

Unused here, but present for interface compatiblity with findMsMsHR

rtLimits

Unused here, but present for interface compatiblity with findMsMsHR

maxCount

Maximal number of spectra to return

headerCache

Cached results of header(msRaw), either to speed up the operations or to operate with preselected header() data

fillPrecursorScan

Unused here, but present for interface compatiblity with findMsMsHR

deprofile

Whether deprofiling should take place, and what method should be used (cf. deprofile)

trace

Either "ms2tic" or "ms2basepeak": Which intensity trace to use - can be either the TIC of the MS2 or the basepeak intensity of the MS2.

Value

a list of "spectrum sets" as defined in findMsMsHR, sorted by decreasing precursor intensity.

Details

Note that this is not a precise function and only really makes sense in direct infusion and if the precursor is really known, because MS2 precursor data is only "roughly" accurate (to 2 dp). The regular findMsMsHR functions confirm the exact mass of the precursor in the MS1 scan.

Author

stravsmi