Find compound information

Retrieves compound information from the loaded compound list or calculates it from the SMILES code in the list.

findMz(cpdID, mode = "pH", ppm = 10, deltaMz = 0, retrieval="standard",
unknownMass = getOption("RMassBank")$unknownMass )

findRt(cpdID) 

findSmiles(cpdID) 

findFormula(cpdID, retrieval="standard") 

findCAS(cpdID)

findName(cpdID)

findLevel(cpdID, compact=FALSE)

Arguments

cpdID

The compound ID in the compound list.

mode

Specifies the species of the molecule: An empty string specifies uncharged monoisotopic mass, pH (positive H) specifies [M+H]+, pNa specifies [M+Na]+, pM specifies [M]+, mH and mFA specify [M-H]- and [M+FA]-, respectively. (I apologize for the naming of pH which has absolutely nothing to do with chemical pH values.)

ppm

Specifies ppm window (10 ppm will return the range of the molecular mass + and - 10 ppm).

deltaMz

Specifies additional m/z window to add to the range (deltaMz = 0.02 will return the range of the molecular mass +- 0.02 (and additionally +- the set ppm value).

retrieval

A value that determines whether the files should be handled either as "standard", if the compoundlist is complete, "tentative", if at least a formula is present or "unknown" if the only know thing is the m/z

unknownMass

`charged` or `neutral` (`charged` assumed by default) specifies whether a mass of an unknown compound (level 5) refers to the charged or neutral mass (and correspondingly, whether it must be shifted or not to find the m/z value)

compact

Only for findLevel, returns the "retrieval" parameter used for many functions within RMassBank if TRUE

Value

findMz will return a list(mzCenter=, mzMin=, mzMax=)

with the molecular weight of the given ion, as calculated from the SMILES code and Rcdk.

findRt, findSmiles,findCAS,findName will return the corresponding entry from the compound list. findFormula returns the molecular formula as determined from the SMILES code.

See also

Author

Michael Stravs

Examples