Retrieve annotation data

Retrieves annotation data for an unknown compound by using basic information present

gatherDataUnknown(id, mode, retrieval)

Arguments

id

The compound ID.

mode

"pH", "pNa", "pM", "pNH4", "mH", "mM", "mFA" for different ions ([M+H]+, [M+Na]+, [M]+, [M+NH4]+, [M-H]-, [M]-, [M+FA]-).

retrieval

A value that determines whether the files should be handled either as "standard", if the compoundlist is complete, "tentative", if at least a formula is present or "unknown" if the only know thing is the m/z

Value

Returns a list of type list(id= compoundID, ..., 'ACCESSION' = '', 'RECORD_TITLE' = '', ) etc.

Details

Composes the "upper part" of a MassBank record filled with chemical data about the compound: name, exact mass, structure, CAS no.. The instrument type is also written into this block (even if not strictly part of the chemical information). Additionally, index fields are added at the start of the record, which will be removed later: id, dbcas, dbname from the compound list.

Additionally, the fields ACCESSION and RECORD_TITLE are inserted empty and will be filled later on.

This function is used to generate the data in case a substance is unknown, i.e. not enough information is present to derive anything about formulas or links

References

MassBank record format: http://www.massbank.jp/manuals/MassBankRecord_en.pdf

See also

Author

Michael Stravs, Erik Mueller

Examples

# Gather data for compound ID 131
if (FALSE) gatherDataUnknown(131,"pH")