Create Rcdk molecule from SMILES

Generates a Rcdk molecule object from SMILES code, which is fully typed and usable (in contrast to the built-in parse.smiles).

getMolecule(smiles)

Arguments

smiles

The SMILES code of the compound.

Value

A Rcdk IAtomContainer reference.

Details

NOTE: As of today (2012-03-16), Rcdk discards stereochemistry when loading the SMILES code! Therefore, do not trust this function blindly, e.g. don't generate InChI keys from the result. It is, however, useful if you want to compute the mass (or something else) with Rcdk.

See also

Author

Michael Stravs

Examples

# Lindane:
getMolecule("C1(C(C(C(C(C1Cl)Cl)Cl)Cl)Cl)Cl")
#> [1] "Java-Object{AtomContainer(1209669119, #A:18, AtomRef{Atom(775931202, S:C, H:0, AtomType(775931202, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(775931202, Element(775931202, S:C, AN:6))))}, AtomRef{Atom(22069592, S:C, H:0, AtomType(22069592, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(22069592, Element(22069592, S:C, AN:6))))}, AtomRef{Atom(1160003871, S:C, H:0, AtomType(1160003871, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(1160003871, Element(1160003871, S:C, AN:6))))}, AtomRef{Atom(1075738627, S:C, H:0, AtomType(1075738627, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(1075738627, Element(1075738627, S:C, AN:6))))}, AtomRef{Atom(282828951, S:C, H:0, AtomType(282828951, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(282828951, Element(282828951, S:C, AN:6))))}, AtomRef{Atom(394721749, S:C, H:0, AtomType(394721749, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(394721749, Element(394721749, S:C, AN:6))))}, AtomRef{Atom(1884122755, S:Cl, H:0, AtomType(1884122755, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(1884122755, Element(1884122755, S:Cl, AN:17))))}, AtomRef{Atom(1134612201, S:Cl, H:0, AtomType(1134612201, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(1134612201, Element(1134612201, S:Cl, AN:17))))}, AtomRef{Atom(246550802, S:Cl, H:0, AtomType(246550802, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(246550802, Element(246550802, S:Cl, AN:17))))}, AtomRef{Atom(786041152, S:Cl, H:0, AtomType(786041152, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(786041152, Element(786041152, S:Cl, AN:17))))}, AtomRef{Atom(897074030, S:Cl, H:0, AtomType(897074030, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(897074030, Element(897074030, S:Cl, AN:17))))}, AtomRef{Atom(1885996206, S:Cl, H:0, AtomType(1885996206, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(1885996206, Element(1885996206, S:Cl, AN:17))))}, AtomRef{Atom(1543974463, S:H, H:0, AtomType(1543974463, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(1543974463, Element(1543974463, S:H, AN:1))))}, AtomRef{Atom(1293618474, S:H, H:0, AtomType(1293618474, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(1293618474, Element(1293618474, S:H, AN:1))))}, AtomRef{Atom(156545103, S:H, H:0, AtomType(156545103, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(156545103, Element(156545103, S:H, AN:1))))}, AtomRef{Atom(345281752, S:H, H:0, AtomType(345281752, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(345281752, Element(345281752, S:H, AN:1))))}, AtomRef{Atom(1896294051, S:H, H:0, AtomType(1896294051, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(1896294051, Element(1896294051, S:H, AN:1))))}, AtomRef{Atom(1684015092, S:H, H:0, AtomType(1684015092, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(1684015092, Element(1684015092, S:H, AN:1))))}, #B:18, BondRef{Bond(1920467934, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(775931202, S:C, H:0, AtomType(775931202, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(775931202, Element(775931202, S:C, AN:6))))}, AtomRef{Atom(22069592, S:C, H:0, AtomType(22069592, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(22069592, Element(22069592, S:C, AN:6))))}, ElectronContainer(1920467934EC:2))}, BondRef{Bond(1883840933, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(22069592, S:C, H:0, AtomType(22069592, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(22069592, Element(22069592, S:C, AN:6))))}, AtomRef{Atom(1160003871, S:C, H:0, AtomType(1160003871, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(1160003871, Element(1160003871, S:C, AN:6))))}, ElectronContainer(1883840933EC:2))}, BondRef{Bond(233996206, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(1160003871, S:C, H:0, AtomType(1160003871, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(1160003871, Element(1160003871, S:C, AN:6))))}, AtomRef{Atom(1075738627, S:C, H:0, AtomType(1075738627, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(1075738627, Element(1075738627, S:C, AN:6))))}, ElectronContainer(233996206EC:2))}, BondRef{Bond(614685048, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(1075738627, S:C, H:0, AtomType(1075738627, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(1075738627, Element(1075738627, S:C, AN:6))))}, AtomRef{Atom(282828951, S:C, H:0, AtomType(282828951, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(282828951, Element(282828951, S:C, AN:6))))}, ElectronContainer(614685048EC:2))}, BondRef{Bond(385337537, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(282828951, S:C, H:0, AtomType(282828951, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(282828951, Element(282828951, S:C, AN:6))))}, AtomRef{Atom(394721749, S:C, H:0, AtomType(394721749, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(394721749, Element(394721749, S:C, AN:6))))}, ElectronContainer(385337537EC:2))}, BondRef{Bond(789219251, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(775931202, S:C, H:0, AtomType(775931202, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(775931202, Element(775931202, S:C, AN:6))))}, AtomRef{Atom(394721749, S:C, H:0, AtomType(394721749, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(394721749, Element(394721749, S:C, AN:6))))}, ElectronContainer(789219251EC:2))}, BondRef{Bond(832279283, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(394721749, S:C, H:0, AtomType(394721749, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(394721749, Element(394721749, S:C, AN:6))))}, AtomRef{Atom(1884122755, S:Cl, H:0, AtomType(1884122755, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(1884122755, Element(1884122755, S:Cl, AN:17))))}, ElectronContainer(832279283EC:2))}, BondRef{Bond(265119009, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(282828951, S:C, H:0, AtomType(282828951, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(282828951, Element(282828951, S:C, AN:6))))}, AtomRef{Atom(1134612201, S:Cl, H:0, AtomType(1134612201, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(1134612201, Element(1134612201, S:Cl, AN:17))))}, ElectronContainer(265119009EC:2))}, BondRef{Bond(668210649, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(1075738627, S:C, H:0, AtomType(1075738627, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(1075738627, Element(1075738627, S:C, AN:6))))}, AtomRef{Atom(246550802, S:Cl, H:0, AtomType(246550802, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(246550802, Element(246550802, S:Cl, AN:17))))}, ElectronContainer(668210649EC:2))}, BondRef{Bond(1545087375, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(1160003871, S:C, H:0, AtomType(1160003871, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(1160003871, Element(1160003871, S:C, AN:6))))}, AtomRef{Atom(786041152, S:Cl, H:0, AtomType(786041152, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(786041152, Element(786041152, S:Cl, AN:17))))}, ElectronContainer(1545087375EC:2))}, BondRef{Bond(838411509, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(22069592, S:C, H:0, AtomType(22069592, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(22069592, Element(22069592, S:C, AN:6))))}, AtomRef{Atom(897074030, S:Cl, H:0, AtomType(897074030, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(897074030, Element(897074030, S:Cl, AN:17))))}, ElectronContainer(838411509EC:2))}, BondRef{Bond(1434041222, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(775931202, S:C, H:0, AtomType(775931202, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(775931202, Element(775931202, S:C, AN:6))))}, AtomRef{Atom(1885996206, S:Cl, H:0, AtomType(1885996206, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(1885996206, Element(1885996206, S:Cl, AN:17))))}, ElectronContainer(1434041222EC:2))}, BondRef{Bond(1375995437, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(775931202, S:C, H:0, AtomType(775931202, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(775931202, Element(775931202, S:C, AN:6))))}, Atom(1543974463, S:H, H:0, AtomType(1543974463, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(1543974463, Element(1543974463, S:H, AN:1)))), ElectronContainer(1375995437EC:2))}, BondRef{Bond(1338841523, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(22069592, S:C, H:0, AtomType(22069592, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(22069592, Element(22069592, S:C, AN:6))))}, Atom(1293618474, S:H, H:0, AtomType(1293618474, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(1293618474, Element(1293618474, S:H, AN:1)))), ElectronContainer(1338841523EC:2))}, BondRef{Bond(929776179, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(1160003871, S:C, H:0, AtomType(1160003871, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(1160003871, Element(1160003871, S:C, AN:6))))}, Atom(156545103, S:H, H:0, AtomType(156545103, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(156545103, Element(156545103, S:H, AN:1)))), ElectronContainer(929776179EC:2))}, BondRef{Bond(802581203, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(1075738627, S:C, H:0, AtomType(1075738627, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(1075738627, Element(1075738627, S:C, AN:6))))}, Atom(345281752, S:H, H:0, AtomType(345281752, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(345281752, Element(345281752, S:H, AN:1)))), ElectronContainer(802581203EC:2))}, BondRef{Bond(1561408618, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(282828951, S:C, H:0, AtomType(282828951, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(282828951, Element(282828951, S:C, AN:6))))}, Atom(1896294051, S:H, H:0, AtomType(1896294051, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(1896294051, Element(1896294051, S:H, AN:1)))), ElectronContainer(1561408618EC:2))}, BondRef{Bond(2050404090, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(394721749, S:C, H:0, AtomType(394721749, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(394721749, Element(394721749, S:C, AN:6))))}, Atom(1684015092, S:H, H:0, AtomType(1684015092, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(1684015092, Element(1684015092, S:H, AN:1)))), ElectronContainer(2050404090EC:2))})}"
# Benzene:
getMolecule("C1=CC=CC=C1")
#> [1] "Java-Object{AtomContainer(1256728724, #A:12, AtomRef{Atom(388043093, S:C, H:0, AtomType(388043093, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(388043093, Element(388043093, S:C, AN:6))))}, AtomRef{Atom(188576144, S:C, H:0, AtomType(188576144, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(188576144, Element(188576144, S:C, AN:6))))}, AtomRef{Atom(1608230649, S:C, H:0, AtomType(1608230649, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(1608230649, Element(1608230649, S:C, AN:6))))}, AtomRef{Atom(282432134, S:C, H:0, AtomType(282432134, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(282432134, Element(282432134, S:C, AN:6))))}, AtomRef{Atom(266437232, S:C, H:0, AtomType(266437232, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(266437232, Element(266437232, S:C, AN:6))))}, AtomRef{Atom(1873859565, S:C, H:0, AtomType(1873859565, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(1873859565, Element(1873859565, S:C, AN:6))))}, AtomRef{Atom(673068808, S:H, H:0, AtomType(673068808, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(673068808, Element(673068808, S:H, AN:1))))}, AtomRef{Atom(900008524, S:H, H:0, AtomType(900008524, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(900008524, Element(900008524, S:H, AN:1))))}, AtomRef{Atom(520232556, S:H, H:0, AtomType(520232556, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(520232556, Element(520232556, S:H, AN:1))))}, AtomRef{Atom(17037394, S:H, H:0, AtomType(17037394, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(17037394, Element(17037394, S:H, AN:1))))}, AtomRef{Atom(1484531981, S:H, H:0, AtomType(1484531981, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(1484531981, Element(1484531981, S:H, AN:1))))}, AtomRef{Atom(1159114532, S:H, H:0, AtomType(1159114532, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(1159114532, Element(1159114532, S:H, AN:1))))}, #B:12, BondRef{Bond(1843289228, #O:DOUBLE, #S:NONE, #A:2, AtomRef{Atom(388043093, S:C, H:0, AtomType(388043093, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(388043093, Element(388043093, S:C, AN:6))))}, AtomRef{Atom(188576144, S:C, H:0, AtomType(188576144, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(188576144, Element(188576144, S:C, AN:6))))}, ElectronContainer(1843289228EC:4))}, BondRef{Bond(1361289747, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(188576144, S:C, H:0, AtomType(188576144, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(188576144, Element(188576144, S:C, AN:6))))}, AtomRef{Atom(1608230649, S:C, H:0, AtomType(1608230649, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(1608230649, Element(1608230649, S:C, AN:6))))}, ElectronContainer(1361289747EC:2))}, BondRef{Bond(1381128261, #O:DOUBLE, #S:NONE, #A:2, AtomRef{Atom(1608230649, S:C, H:0, AtomType(1608230649, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(1608230649, Element(1608230649, S:C, AN:6))))}, AtomRef{Atom(282432134, S:C, H:0, AtomType(282432134, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(282432134, Element(282432134, S:C, AN:6))))}, ElectronContainer(1381128261EC:4))}, BondRef{Bond(999609945, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(282432134, S:C, H:0, AtomType(282432134, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(282432134, Element(282432134, S:C, AN:6))))}, AtomRef{Atom(266437232, S:C, H:0, AtomType(266437232, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(266437232, Element(266437232, S:C, AN:6))))}, ElectronContainer(999609945EC:2))}, BondRef{Bond(615634843, #O:DOUBLE, #S:NONE, #A:2, AtomRef{Atom(266437232, S:C, H:0, AtomType(266437232, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(266437232, Element(266437232, S:C, AN:6))))}, AtomRef{Atom(1873859565, S:C, H:0, AtomType(1873859565, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(1873859565, Element(1873859565, S:C, AN:6))))}, ElectronContainer(615634843EC:4))}, BondRef{Bond(1758386724, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(388043093, S:C, H:0, AtomType(388043093, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(388043093, Element(388043093, S:C, AN:6))))}, AtomRef{Atom(1873859565, S:C, H:0, AtomType(1873859565, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(1873859565, Element(1873859565, S:C, AN:6))))}, ElectronContainer(1758386724EC:2))}, BondRef{Bond(1412925683, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(388043093, S:C, H:0, AtomType(388043093, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(388043093, Element(388043093, S:C, AN:6))))}, Atom(673068808, S:H, H:0, AtomType(673068808, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(673068808, Element(673068808, S:H, AN:1)))), ElectronContainer(1412925683EC:2))}, BondRef{Bond(1832580921, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(188576144, S:C, H:0, AtomType(188576144, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(188576144, Element(188576144, S:C, AN:6))))}, Atom(900008524, S:H, H:0, AtomType(900008524, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(900008524, Element(900008524, S:H, AN:1)))), ElectronContainer(1832580921EC:2))}, BondRef{Bond(497359413, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(1608230649, S:C, H:0, AtomType(1608230649, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(1608230649, Element(1608230649, S:C, AN:6))))}, Atom(520232556, S:H, H:0, AtomType(520232556, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(520232556, Element(520232556, S:H, AN:1)))), ElectronContainer(497359413EC:2))}, BondRef{Bond(369241501, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(282432134, S:C, H:0, AtomType(282432134, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(282432134, Element(282432134, S:C, AN:6))))}, Atom(17037394, S:H, H:0, AtomType(17037394, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(17037394, Element(17037394, S:H, AN:1)))), ElectronContainer(369241501EC:2))}, BondRef{Bond(2124046270, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(266437232, S:C, H:0, AtomType(266437232, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(266437232, Element(266437232, S:C, AN:6))))}, Atom(1484531981, S:H, H:0, AtomType(1484531981, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(1484531981, Element(1484531981, S:H, AN:1)))), ElectronContainer(2124046270EC:2))}, BondRef{Bond(1151593579, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(1873859565, S:C, H:0, AtomType(1873859565, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(1873859565, Element(1873859565, S:C, AN:6))))}, Atom(1159114532, S:H, H:0, AtomType(1159114532, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(1159114532, Element(1159114532, S:H, AN:1)))), ElectronContainer(1151593579EC:2))})}"