Merge peaks for spectra merging, FT shoulder elimination etc.

This procedure first sorts peaks by intensity (descending sort) and then starts iterating over the peaks, removing all entries that deviate "sufficiently far" from the currently selected peak. See the Details section for a full explanation and information on how to fine-tune peak removal.

mergePeaks(peaks, ...)

# S4 method for class 'data.frame'
mergePeaks(peaks, ...)

# S4 method for class 'matrix'
mergePeaks(peaks, ...)

# S4 method for class 'RmbSpectrum2'
mergePeaks(peaks, ...)

# S4 method for class 'Spectrum'
mergePeaks(peaks, ...)

Arguments

peaks

data.frame, matrix or RmbSpectrum2 The peak-table to be merged. In case of an RmbSpectrum2-object, peaks are retrieved and updated via getData and setData, respectively

...

3 numeric values These define cutoff limits (see details)

Value

object of the same class as peaks The result contains a reduced peak-table ordered by m/z

Details

Three parameters must be passed to mergePeaks for peak-removal control in this order: - cutoff_dppm_limit - cutoff_absolute_limit - cutoff_intensity_limit The method iterates through the peaks, beginning with the highest-intensity peak and in each step removes all other peaks that fulfill conditions 1 AND 2 relative to the selected peak 1. Their m/z value does not deviate too far from the one of the selected peak. i.e. if the selected peak is p and the checked peak is c, it holds that EITHER |p$mz - c$mz| <= cutoff_absolute_limit OR |p$mz - c$mz| <= ppm(p$mz, cutoff_dppm_limit, p=TRUE) (see ppm) 2. Their intensity is much smaller than the one of the selected peak, i.e. c$mz < cutoff_intensity_limit * p$mz for a suitable cutoff_intensity_limit between 0 and 1.

See also

Examples

if (FALSE) mergePeaks(spectrum, 10, 0.5, 0.05)