RMassBank mass spectrometry pipeline

Extracts and processes spectra from a specified file list, according to loaded options and given parameters.

msmsWorkflow(
  w,
  mode = "pH",
  steps = c(1:8),
  confirmMode = FALSE,
  newRecalibration = TRUE,
  useRtLimit = TRUE,
  archivename = NA,
  readMethod = "mzR",
  filetable = NULL,
  findPeaksArgs = NULL,
  plots = FALSE,
  precursorscan.cf = FALSE,
  settings = getOption("RMassBank"),
  analyzeMethod = "formula",
  progressbar = "progressBarHook",
  MSe = FALSE
)

Arguments

w

A msmsWorkspace to work with.

mode

"pH", "pNa", "pM", "mH", "mM", "mFA", "pNH4" for different ions ([M+H]+, [M+Na]+, [M]+, [M-H]-, [M]-, [M+FA]-, [M+NH4]+).

steps

Which steps of the workflow to process. See the vignette vignette("RMassBank") for details.

confirmMode

Defaults to false (use most intense precursor). Value 1 uses the 2nd-most intense precursor for a chosen ion (and its data-dependent scans) , etc.

newRecalibration

Whether to generate a new recalibration curve (TRUE, default) or to reuse the currently stored curve (FALSE, useful e.g. for adduct-processing runs.)

useRtLimit

Whether to enforce the given retention time window.

archivename

The prefix under which to store the analyzed result files.

readMethod

Several methods are available to get peak lists from the files. Currently supported are "mzR", "xcms", "MassBank" and "peaklist". The first two read MS/MS raw data, and differ in the strategy used to extract peaks. MassBank will read existing records, so that e.g. a recalibration can be performed, and "peaklist" just requires a CSV with two columns and the column header "mz", "int".

filetable

If including step 1 (data extraction), a `filetable` argument to be passed to msmsRead.

findPeaksArgs

A list of arguments that will be handed to the xcms-method findPeaks via do.call

plots

A parameter that determines whether the spectra should be plotted or not (This parameter is only used for the xcms-method)

precursorscan.cf

Whether to fill precursor scans. To be used with files which for some reasons do not contain precursor scan IDs in the mzML, e.g. AB Sciex converted files.

settings

Options to be used for processing. Defaults to the options loaded via loadRmbSettings et al. Refer to there for specific settings.

analyzeMethod

The "method" parameter to pass to analyzeMsMs.

progressbar

The progress bar callback to use. Only needed for specialized applications. Cf. the documentation of progressBarHook for usage.

MSe

A boolean value that determines whether the spectra were recorded using MSe or not

Value

The processed msmsWorkspace.

Details

The filenames of the raw LC-MS runs are read from the array files in the global enviroment. See the vignette vignette("RMassBank") for further details about the workflow.

See also

Author

Michael Stravs, Eawag <michael.stravs@eawag.ch>