Identify intense peaks (in a list of unmatched peaks)

Finds a list of peaks in spectra with a high relative intensity (>10 1e4, or >1 checked. Peaks orbiting around the parent peak mass (calculated from the compound ID), which are very likely co-isolated substances, are ignored.

problematicPeaks(sp)

Arguments

sp

a RmbSpectrum2 object to be checked for problematic peaks.

Value

The modified RmbSpectrum2 object with additional columns/properties `problematicPeaks` (logical `TRUE` if the peak is intense and unannotated), `aMax` (base peak intensity), `mzCenter` (the precursor m/z).

Note

TODO: there is hardcoded logic in this function that needs to be resolved eventually!

See also

Author

Michael Stravs

Examples

# As used in the workflow: 

sp <- new("RmbSpectrum2", mz = c(100,200,300,400,500), intensity = c(999999,888888,777777,666666,555555))
sp@ok <- TRUE
property(sp, "mzFound", addNew=TRUE) <- sp@mz
sp@good <- c(TRUE, TRUE, TRUE, FALSE, FALSE)
sp@precursorMz <- 600
sp_checked <- problematicPeaks(sp)
# stopifnot(sum(getData(sp_checked)$problematicPeak) == 2)