Reanalyze unmatched peaks

Reanalysis of peaks with no matching molecular formula by allowing additional elements (e.g. "N2O").

reanalyzeFailpeaks(w, custom_additions, filterSettings =
        getOption("RMassBank")$filterSettings, progressbar = "progressBarHook")
reanalyzeFailpeak(mass, custom_additions, cpdID, mode,
        filterSettings = getOption("RMassBank")$filterSettings)

Arguments

w

A `msmsWorkspace` with annotated peaks.

custom_additions

The allowed additions, e.g. "N2O".

filterSettings

Settings for filtering data. Refer toanalyzeMsMs for settings.

progressbar

The progress bar callback to use. Only needed for specialized applications. Cf. the documentation of progressBarHook for usage.

mass

(Usually recalibrated) m/z value of the peak.

cpdID

Compound ID of this spectrum.

mode

for `reanalyzeFailpeak`, the `mode` (adduct) of the analyzed spectrum.

Value

The aggregate data frame extended by the columns: #'

reanalyzed.???

If reanalysis (step 7) has already been processed: matching values from the reanalyzed peaks

matchedReanalysis

Whether reanalysis has matched (TRUE), not matched(FALSE) or has not been conducted for the peak(NA).

It would be good to merge the analysis functions of analyzeMsMs with the one used here, to simplify code changes.

Details

reanalyzeFailpeaks examines the unmatchedPeaksC table in specs and sends every peak through reanalyzeFailpeak.

See also

Author

Michael Stravs

Examples

## As used in the workflow:
if (FALSE) {    
  reanalyzedRcSpecs <- reanalyzeFailpeaks(w@aggregated, custom_additions="N2O", mode="pH")
# A single peak:
reanalyzeFailpeak("N2O", 105.0447, 1234, 1, 1)
}