Write MassBank record into character array

Writes a MassBank record in list format to a text array.

toMassbank(o, ...)

# S4 method for class 'RmbSpectraSet'
toMassbank(o, addAnnotation = getOption("RMassBank")$add_annotation)

# S4 method for class 'RmbSpectrum2'
toMassbank(o, addAnnotation = getOption("RMassBank")$add_annotation)

Arguments

o

An object to convert to MassBank record format. This may be a single `RmbSpectrum2`, or a complete compound (an `RmbSpectraSet`),

...

Parameters passed to the implementation, in particular `addAnnotation`

addAnnotation

`logical`, whether to add peak annotations (putative formulas) to the record.

Value

The result is a text array, which is ready to be written to the disk as a file.

Details

The function is a general conversion tool for the MassBank format; i.e. the field names are not fixed. mbdata must be a named list, and the entries can be as follows:

A single text line:

'CH\$EXACT_MASS' = '329.1023'

is written as

CH\$EXACT_MASS: 329.1023

A character array:

'CH\$NAME' = c('2-Aminobenzimidazole', '1H-Benzimidazol-2-amine')

is written as

CH\$NAME: 2-Aminobenzimidazole

CH\$NAME: 1H-Benzimidazol-2-amine

A named list of strings:

'CH\$LINK' = list('CHEBI' = "27822", "KEGG" = "C10901")

is written as

CH\$LINK: CHEBI 27822

CH\$LINK: KEGG C10901

A data frame (e.g. the peak table) is written as specified in the MassBank record format (Section 2.6.3): the column names are used as headers for the first line, all data rows are printed space-separated.

Note

The function iterates over the list item names. This means that duplicate entries in mbdata are (partially) discarded! The correct way to add them is by making a character array (as specified above): Instead of 'CH\$NAME' = 'bla', 'CH\$NAME' = 'blub' specify 'CH\$NAME' = c('bla','blub').

References

MassBank record format: http://www.massbank.jp/manuals/MassBankRecord_en.pdf

See also

Author

Michael Stravs

Examples

if (FALSE) {
# Read just the compound info skeleton from the Internet for some compound ID
id <- 35
mbdata <- gatherData(id)
#' # Export the mbdata blocks to line arrays 
# (there is no spectrum information, just the compound info...)
mbtext <- toMassbank(mbdata)
}