Create Rcdk molecule from SMILES

Generates a Rcdk molecule object from SMILES code, which is fully typed and usable (in contrast to the built-in parse.smiles).

getMolecule(smiles)

Arguments

smiles

The SMILES code of the compound.

Value

A Rcdk IAtomContainer reference.

Details

NOTE: As of today (2012-03-16), Rcdk discards stereochemistry when loading the SMILES code! Therefore, do not trust this function blindly, e.g. don't generate InChI keys from the result. It is, however, useful if you want to compute the mass (or something else) with Rcdk.

See also

Author

Michael Stravs

Examples

# Lindane:
getMolecule("C1(C(C(C(C(C1Cl)Cl)Cl)Cl)Cl)Cl")
#> [1] "Java-Object{AtomContainer(1742920067, #A:18, AtomRef{Atom(1663619914, S:C, H:0, AtomType(1663619914, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(1663619914, Element(1663619914, S:C, AN:6))))}, AtomRef{Atom(341748265, S:C, H:0, AtomType(341748265, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(341748265, Element(341748265, S:C, AN:6))))}, AtomRef{Atom(364604394, S:C, H:0, AtomType(364604394, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(364604394, Element(364604394, S:C, AN:6))))}, AtomRef{Atom(146370526, S:C, H:0, AtomType(146370526, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(146370526, Element(146370526, S:C, AN:6))))}, AtomRef{Atom(758013696, S:C, H:0, AtomType(758013696, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(758013696, Element(758013696, S:C, AN:6))))}, AtomRef{Atom(1279309678, S:C, H:0, AtomType(1279309678, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(1279309678, Element(1279309678, S:C, AN:6))))}, AtomRef{Atom(48914743, S:Cl, H:0, AtomType(48914743, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(48914743, Element(48914743, S:Cl, AN:17))))}, AtomRef{Atom(1106131243, S:Cl, H:0, AtomType(1106131243, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(1106131243, Element(1106131243, S:Cl, AN:17))))}, AtomRef{Atom(510109769, S:Cl, H:0, AtomType(510109769, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(510109769, Element(510109769, S:Cl, AN:17))))}, AtomRef{Atom(1473611564, S:Cl, H:0, AtomType(1473611564, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(1473611564, Element(1473611564, S:Cl, AN:17))))}, AtomRef{Atom(107456312, S:Cl, H:0, AtomType(107456312, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(107456312, Element(107456312, S:Cl, AN:17))))}, AtomRef{Atom(921760190, S:Cl, H:0, AtomType(921760190, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(921760190, Element(921760190, S:Cl, AN:17))))}, AtomRef{Atom(1074593562, S:H, H:0, AtomType(1074593562, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(1074593562, Element(1074593562, S:H, AN:1))))}, AtomRef{Atom(660017404, S:H, H:0, AtomType(660017404, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(660017404, Element(660017404, S:H, AN:1))))}, AtomRef{Atom(1381965390, S:H, H:0, AtomType(1381965390, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(1381965390, Element(1381965390, S:H, AN:1))))}, AtomRef{Atom(1979313356, S:H, H:0, AtomType(1979313356, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(1979313356, Element(1979313356, S:H, AN:1))))}, AtomRef{Atom(1386883398, S:H, H:0, AtomType(1386883398, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(1386883398, Element(1386883398, S:H, AN:1))))}, AtomRef{Atom(1306854175, S:H, H:0, AtomType(1306854175, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(1306854175, Element(1306854175, S:H, AN:1))))}, #B:18, BondRef{Bond(1173643169, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(1663619914, S:C, H:0, AtomType(1663619914, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(1663619914, Element(1663619914, S:C, AN:6))))}, AtomRef{Atom(341748265, S:C, H:0, AtomType(341748265, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(341748265, Element(341748265, S:C, AN:6))))}, ElectronContainer(1173643169EC:2))}, BondRef{Bond(1282287470, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(341748265, S:C, H:0, AtomType(341748265, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(341748265, Element(341748265, S:C, AN:6))))}, AtomRef{Atom(364604394, S:C, H:0, AtomType(364604394, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(364604394, Element(364604394, S:C, AN:6))))}, ElectronContainer(1282287470EC:2))}, BondRef{Bond(1397616978, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(364604394, S:C, H:0, AtomType(364604394, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(364604394, Element(364604394, S:C, AN:6))))}, AtomRef{Atom(146370526, S:C, H:0, AtomType(146370526, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(146370526, Element(146370526, S:C, AN:6))))}, ElectronContainer(1397616978EC:2))}, BondRef{Bond(1390835631, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(146370526, S:C, H:0, AtomType(146370526, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(146370526, Element(146370526, S:C, AN:6))))}, AtomRef{Atom(758013696, S:C, H:0, AtomType(758013696, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(758013696, Element(758013696, S:C, AN:6))))}, ElectronContainer(1390835631EC:2))}, BondRef{Bond(889729797, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(758013696, S:C, H:0, AtomType(758013696, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(758013696, Element(758013696, S:C, AN:6))))}, AtomRef{Atom(1279309678, S:C, H:0, AtomType(1279309678, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(1279309678, Element(1279309678, S:C, AN:6))))}, ElectronContainer(889729797EC:2))}, BondRef{Bond(148912029, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(1663619914, S:C, H:0, AtomType(1663619914, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(1663619914, Element(1663619914, S:C, AN:6))))}, AtomRef{Atom(1279309678, S:C, H:0, AtomType(1279309678, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(1279309678, Element(1279309678, S:C, AN:6))))}, ElectronContainer(148912029EC:2))}, BondRef{Bond(874217650, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(1279309678, S:C, H:0, AtomType(1279309678, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(1279309678, Element(1279309678, S:C, AN:6))))}, AtomRef{Atom(48914743, S:Cl, H:0, AtomType(48914743, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(48914743, Element(48914743, S:Cl, AN:17))))}, ElectronContainer(874217650EC:2))}, BondRef{Bond(1436664465, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(758013696, S:C, H:0, AtomType(758013696, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(758013696, Element(758013696, S:C, AN:6))))}, AtomRef{Atom(1106131243, S:Cl, H:0, AtomType(1106131243, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(1106131243, Element(1106131243, S:Cl, AN:17))))}, ElectronContainer(1436664465EC:2))}, BondRef{Bond(558187323, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(146370526, S:C, H:0, AtomType(146370526, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(146370526, Element(146370526, S:C, AN:6))))}, AtomRef{Atom(510109769, S:Cl, H:0, AtomType(510109769, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(510109769, Element(510109769, S:Cl, AN:17))))}, ElectronContainer(558187323EC:2))}, BondRef{Bond(680576081, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(364604394, S:C, H:0, AtomType(364604394, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(364604394, Element(364604394, S:C, AN:6))))}, AtomRef{Atom(1473611564, S:Cl, H:0, AtomType(1473611564, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(1473611564, Element(1473611564, S:Cl, AN:17))))}, ElectronContainer(680576081EC:2))}, BondRef{Bond(1088872417, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(341748265, S:C, H:0, AtomType(341748265, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(341748265, Element(341748265, S:C, AN:6))))}, AtomRef{Atom(107456312, S:Cl, H:0, AtomType(107456312, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(107456312, Element(107456312, S:Cl, AN:17))))}, ElectronContainer(1088872417EC:2))}, BondRef{Bond(453523494, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(1663619914, S:C, H:0, AtomType(1663619914, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(1663619914, Element(1663619914, S:C, AN:6))))}, AtomRef{Atom(921760190, S:Cl, H:0, AtomType(921760190, N:Cl, MBO:SINGLE, BOS:1.0, FC:0, H:SP3, NC:1, EV:1, Isotope(921760190, Element(921760190, S:Cl, AN:17))))}, ElectronContainer(453523494EC:2))}, BondRef{Bond(274722023, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(1663619914, S:C, H:0, AtomType(1663619914, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(1663619914, Element(1663619914, S:C, AN:6))))}, Atom(1074593562, S:H, H:0, AtomType(1074593562, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(1074593562, Element(1074593562, S:H, AN:1)))), ElectronContainer(274722023EC:2))}, BondRef{Bond(1052967153, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(341748265, S:C, H:0, AtomType(341748265, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(341748265, Element(341748265, S:C, AN:6))))}, Atom(660017404, S:H, H:0, AtomType(660017404, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(660017404, Element(660017404, S:H, AN:1)))), ElectronContainer(1052967153EC:2))}, BondRef{Bond(1210898719, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(364604394, S:C, H:0, AtomType(364604394, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(364604394, Element(364604394, S:C, AN:6))))}, Atom(1381965390, S:H, H:0, AtomType(1381965390, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(1381965390, Element(1381965390, S:H, AN:1)))), ElectronContainer(1210898719EC:2))}, BondRef{Bond(306123060, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(146370526, S:C, H:0, AtomType(146370526, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(146370526, Element(146370526, S:C, AN:6))))}, Atom(1979313356, S:H, H:0, AtomType(1979313356, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(1979313356, Element(1979313356, S:H, AN:1)))), ElectronContainer(306123060EC:2))}, BondRef{Bond(2104028992, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(758013696, S:C, H:0, AtomType(758013696, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(758013696, Element(758013696, S:C, AN:6))))}, Atom(1386883398, S:H, H:0, AtomType(1386883398, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(1386883398, Element(1386883398, S:H, AN:1)))), ElectronContainer(2104028992EC:2))}, BondRef{Bond(1527430292, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(1279309678, S:C, H:0, AtomType(1279309678, N:C.sp3, MBO:SINGLE, BOS:4.0, FC:0, H:SP3, NC:4, EV:4, Isotope(1279309678, Element(1279309678, S:C, AN:6))))}, Atom(1306854175, S:H, H:0, AtomType(1306854175, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(1306854175, Element(1306854175, S:H, AN:1)))), ElectronContainer(1527430292EC:2))})}"
# Benzene:
getMolecule("C1=CC=CC=C1")
#> [1] "Java-Object{AtomContainer(1418334255, #A:12, AtomRef{Atom(1975546571, S:C, H:0, AtomType(1975546571, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(1975546571, Element(1975546571, S:C, AN:6))))}, AtomRef{Atom(521960438, S:C, H:0, AtomType(521960438, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(521960438, Element(521960438, S:C, AN:6))))}, AtomRef{Atom(632587706, S:C, H:0, AtomType(632587706, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(632587706, Element(632587706, S:C, AN:6))))}, AtomRef{Atom(726950788, S:C, H:0, AtomType(726950788, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(726950788, Element(726950788, S:C, AN:6))))}, AtomRef{Atom(1567885839, S:C, H:0, AtomType(1567885839, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(1567885839, Element(1567885839, S:C, AN:6))))}, AtomRef{Atom(558569884, S:C, H:0, AtomType(558569884, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(558569884, Element(558569884, S:C, AN:6))))}, AtomRef{Atom(631659383, S:H, H:0, AtomType(631659383, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(631659383, Element(631659383, S:H, AN:1))))}, AtomRef{Atom(785447854, S:H, H:0, AtomType(785447854, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(785447854, Element(785447854, S:H, AN:1))))}, AtomRef{Atom(556529265, S:H, H:0, AtomType(556529265, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(556529265, Element(556529265, S:H, AN:1))))}, AtomRef{Atom(346224929, S:H, H:0, AtomType(346224929, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(346224929, Element(346224929, S:H, AN:1))))}, AtomRef{Atom(63468833, S:H, H:0, AtomType(63468833, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(63468833, Element(63468833, S:H, AN:1))))}, AtomRef{Atom(720167805, S:H, H:0, AtomType(720167805, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(720167805, Element(720167805, S:H, AN:1))))}, #B:12, BondRef{Bond(550752602, #O:DOUBLE, #S:NONE, #A:2, AtomRef{Atom(1975546571, S:C, H:0, AtomType(1975546571, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(1975546571, Element(1975546571, S:C, AN:6))))}, AtomRef{Atom(521960438, S:C, H:0, AtomType(521960438, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(521960438, Element(521960438, S:C, AN:6))))}, ElectronContainer(550752602EC:4))}, BondRef{Bond(680779399, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(521960438, S:C, H:0, AtomType(521960438, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(521960438, Element(521960438, S:C, AN:6))))}, AtomRef{Atom(632587706, S:C, H:0, AtomType(632587706, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(632587706, Element(632587706, S:C, AN:6))))}, ElectronContainer(680779399EC:2))}, BondRef{Bond(1439337960, #O:DOUBLE, #S:NONE, #A:2, AtomRef{Atom(632587706, S:C, H:0, AtomType(632587706, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(632587706, Element(632587706, S:C, AN:6))))}, AtomRef{Atom(726950788, S:C, H:0, AtomType(726950788, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(726950788, Element(726950788, S:C, AN:6))))}, ElectronContainer(1439337960EC:4))}, BondRef{Bond(741669172, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(726950788, S:C, H:0, AtomType(726950788, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(726950788, Element(726950788, S:C, AN:6))))}, AtomRef{Atom(1567885839, S:C, H:0, AtomType(1567885839, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(1567885839, Element(1567885839, S:C, AN:6))))}, ElectronContainer(741669172EC:2))}, BondRef{Bond(315860201, #O:DOUBLE, #S:NONE, #A:2, AtomRef{Atom(1567885839, S:C, H:0, AtomType(1567885839, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(1567885839, Element(1567885839, S:C, AN:6))))}, AtomRef{Atom(558569884, S:C, H:0, AtomType(558569884, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(558569884, Element(558569884, S:C, AN:6))))}, ElectronContainer(315860201EC:4))}, BondRef{Bond(604125138, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(1975546571, S:C, H:0, AtomType(1975546571, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(1975546571, Element(1975546571, S:C, AN:6))))}, AtomRef{Atom(558569884, S:C, H:0, AtomType(558569884, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(558569884, Element(558569884, S:C, AN:6))))}, ElectronContainer(604125138EC:2))}, BondRef{Bond(1466073198, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(1975546571, S:C, H:0, AtomType(1975546571, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(1975546571, Element(1975546571, S:C, AN:6))))}, Atom(631659383, S:H, H:0, AtomType(631659383, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(631659383, Element(631659383, S:H, AN:1)))), ElectronContainer(1466073198EC:2))}, BondRef{Bond(398690014, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(521960438, S:C, H:0, AtomType(521960438, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(521960438, Element(521960438, S:C, AN:6))))}, Atom(785447854, S:H, H:0, AtomType(785447854, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(785447854, Element(785447854, S:H, AN:1)))), ElectronContainer(398690014EC:2))}, BondRef{Bond(1526298704, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(632587706, S:C, H:0, AtomType(632587706, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(632587706, Element(632587706, S:C, AN:6))))}, Atom(556529265, S:H, H:0, AtomType(556529265, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(556529265, Element(556529265, S:H, AN:1)))), ElectronContainer(1526298704EC:2))}, BondRef{Bond(1593180232, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(726950788, S:C, H:0, AtomType(726950788, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(726950788, Element(726950788, S:C, AN:6))))}, Atom(346224929, S:H, H:0, AtomType(346224929, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(346224929, Element(346224929, S:H, AN:1)))), ElectronContainer(1593180232EC:2))}, BondRef{Bond(492079624, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(1567885839, S:C, H:0, AtomType(1567885839, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(1567885839, Element(1567885839, S:C, AN:6))))}, Atom(63468833, S:H, H:0, AtomType(63468833, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(63468833, Element(63468833, S:H, AN:1)))), ElectronContainer(492079624EC:2))}, BondRef{Bond(380242442, #O:SINGLE, #S:NONE, #A:2, AtomRef{Atom(558569884, S:C, H:0, AtomType(558569884, N:C.sp2, MBO:DOUBLE, BOS:4.0, FC:0, H:SP2, NC:3, EV:4, Isotope(558569884, Element(558569884, S:C, AN:6))))}, Atom(720167805, S:H, H:0, AtomType(720167805, N:H, MBO:SINGLE, BOS:1.0, FC:0, H:S, NC:1, EV:1, Isotope(720167805, Element(720167805, S:H, AN:1)))), ElectronContainer(380242442EC:2))})}"